requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Database
duckdb>=0.9.0

# Audio Processing
faster-whisper>=0.10.0
yt-dlp>=2023.11.0

# LLM
anthropic>=0.25.0

# Newsletter Processing
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Podcast Processing
feedparser>=6.0.10

# Templating
jinja2>=3.1.0

# Utilities
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
